requirements.txt
requests>=2.25.0
pyyaml>=5.4.0
pycryptodome>=3.15.0